# Removed empty events

Changes:
- The global segment was enabling keyboard callbacks for keys UP and Down but those callbacks were empty. Removed those callbacks.

This is a performance measurement of me walljumping on map 2:
Before:
> eventKeyboard: 109ms / 342calls == 0.318713 ms/call
After:
> eventKeyboard: 75ms / 220calls == 0.340909 ms/call

Because the up key is not bound anymore, it is also called less, so what matters is the total times going from 109ms to 75ms.



# Added specific keyboard down and up callbacks

Changes:
- Segments can now optionally bind key down and key up events separately using a table instead of a function.
- Fixed keybinds not being unbound when a segment is deactivated.

This is a performance measurement of me walljumping on map 2, with default segments enabled:
Before:
> eventKeyboard: 75ms / 220calls == 0.340909 ms/call
After:
> eventKeyboard: 25ms / 118calls == 0.211864 ms/call

This is a performance measurement of me walljumping on map 2, with additional segments `dash`, `draw`, `drawonme`, `ffa`, `speed` enabled:
Before:
> eventKeyboard: 188ms / 538calls == 0.349442 ms/call
After:
> eventKeyboard: 94ms / 264calls == 0.356060 ms/call

The number of events triggered is made lower by the changes, so what matter is the total time.

**/!\ This second change have not been updated in the sources yet because sources does not compile nor match the compiled script**
**Do not merge it if you want me to update the sources, tell me if i should do it before or after you updated them.**